section_c/requirements.txt
flask==3.0.2
werkzeug==3.0.1
python-dotenv==1.0.1
flask-cors==4.0.0
selenium==4.18.1
pandas==2.2.1 